fix(onboarding): rename useVerisons -> useVersions
Fix the long-standing filename typo in the onboarding versioning
helper. Also update the single dashboard consumer's import path and
the onboarding CLAUDE.md note that previously documented the typo.

Pure rename + import path swap; no behavior change in this commit.
